HOCS-5361: rework somu implementation
SOMU form implementation currently runs through the actions mechanic for
retrieving the forms and handling the submission. While this was okay
for the initial development, as tenants are requiring specific
functionality this has become clunky and a pain to work around.

This change reworks SOMU implementations so that forms are retrieved
like 'global' forms outside of a workflow. This gives us the ability to
tailor them to the tenants themselves. Additionally, these forms can
also utilise normal case data attributes for visibility conditions.

This change also cleans up some of the initial SOMU development in the
following aspects:
- All posts go through the same API handling
- Getting an individual SOMU item no longer first retrieves them all
- Cleans up submission to remove unused code