Remove nodecron usage
Nodecron was initially implemented to support with refreshing static
lists that could be updated through the MUI. This implementation never
fully worked and ended up causing DOS attacks by requesting a refresh if
the list is empty (as not atomic).

This change removes nodecron from the frontend as no longer required
and we should look at improving the solution through either a caching
layer or in the backend.