HOCS-5560: rework unallocated accordion
Rework of the unallocated accordion to support showing `somu-list`s.
This works by first analyzing the component and if it is a `somu-list`
it gets the meta-data and then the case items for it. This populates the
custom choice behaviour and then composes the titles for each of the
items.

The data for this unallocated accordion is calculated first which is
then used to filter out fields from the accordion that do not have a
data mapping.