Use en dash instead of hyphen-minus in test banner
The non-production banner uses a hyphen (-) instead of a dash (–), which
is incorrect grammatically and does not match the production banner,
which uses an en dash.

This commit changes the hyphen in that banner to use an en dash, which
is UK usage. Some sources say to use an em dash instead, but that
appears limited to US English.