Correct contribution request validation
Validation on contributions are currently trying to parse falsey values,
this change ensures we only call `JSON.parse` on values that aren't
falsey.

This change also simplifies the logic to default invalid and calculate
for a true value.

Exceptions will not only be thrown if the format of the JSON is
incorrect.