HOCS-5615: rework Workstack form submission
The current use of FormData and `append` results in undefined values being passed with the value as string "undefined".

This is causing parsing errors within the Node server and further downstream as "undefined" is not valid JSON.

This change reworks this FormData propagation to strip out any underlying null/undefined.