HOCS-4931: improve skip link
This change pulls out the skip-link into its own distinct class to
better utilise React. This new class handles the scrolling to the
`main-content` id if it exists. To enable tab indexes to work after the
focus, a value of "-1" has been provided. This makes it so that
programmatically it can be focused but not through a keyboard.

This change should also fix HOCS-2050.