Improve correlation ID use
When requesting to backend services, there are currently places where we
don't propagate the request id down to use as a correlation id. This is
causing the inability to trace most logs from the frontend downstream.

Alongside this, we aren't passing the request id into the logger
resulting in instances were we cannot correlate even within the service.