HOCS-6562: change document handling
Previously we're using the casework service as a permission middleware
and sending the entire document through this service. This results in
high memory usage and utilisation through this service, exaggerating
existing performance issues.

This change adds a permission specific check that doesn't passthrough
the entire file. After this permission check passes, the document is
then requested from the document service directly.