HOCS-6656: improve date handling
Change the way that dates are handled in the process middleware. This
now uses regex matching to retrieve the day, month, and year. These
regexes ensure that if a non-alphanumeric character is found either side
of the numbers, then the result of the regex is an empty value for the
specific value. Each of the values allow for leading 0's however, these
are then stripped from the result and then for day and month re-padded
to length 2.